refactor: 인터페이스 리팩토링
