+++ frontend/src/features/auth/types/user.types.ts
@@ -1,0 +1,7 @@
+/**
+ * 사용자 인터페이스
+ */
+export type User = {
+    userId: number;
+    nickname: string;
+};
